pyproject.toml
[project]
name = "reranking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-text-splitters>=0.3.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "pinecone>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "cohere>=5.0.0",
]
